# PERF: Improve perf of to_datetime with ISO format