Addresses #10178

Using the example data from the issue:

    In [2]: df = DataFrame({'date_text':["2015-05-18" for i in range(10**6)]})


Before:

    In [3]: %timeit pd.to_datetime(df['date_text'],format="%Y-%m-%d", box=False).values.view('i8')/10**9
    1 loops, best of 3: 3.14 s per loop
    
    In [4]: %timeit pd.to_datetime(df['date_text'],infer_datetime_format=True,     box=False).values.view('i8')/10**9
    1 loops, best of 3: 253 ms per lo

After:

    In [6]: %timeit pd.to_datetime(df['date_text'],format="%Y-%m-%d", box=False).values.view('i8')/10**9
    1 loops, best of 3: 217 ms per loop
    
    In [7]: %timeit pd.to_datetime(df['date_text'],infer_datetime_format=True, box=False).values.view('i8')/10**9
    1 loops, best of 3: 243 ms per loop