@chris-b1 2015-07-18
I added a vbench based on existing one that inferred the iso860 format.

```
---------------------------------------------------------
Test name                                    |    #0    |
---------------------------------------------------------
timeseries_to_datetime_iso8601_format        |   4.9593 |
timeseries_to_datetime_YYYYMMDD              |  10.0046 |
timeseries_to_datetime_iso8601               |   4.9581 |
---------------------------------------------------------
Test name                                    |    #0    |
---------------------------------------------------------
```

The second line of this test is failing - is that really the desired behavior?  The comments reference #10154  - Wasn't the conclusion that `errors='ignore'` should always suppress?

    def test_day_not_in_month_coerce_false_ignore(self):
        self.assertEqual(to_datetime('2015-02-29', errors='ignore', coerce=False), '2015-02-29')
        self.assertRaises(ValueError, to_datetime, '2015-02-29', errors='ignore', format="%Y-%m-%d", coerce=False)
        self.assertRaises(ValueError, to_datetime, '2015-02-32', errors='ignore', format="%Y-%m-%d", coerce=False)
        self.assertRaises(ValueError, to_datetime, '2015-04-31', errors='ignore', format="%Y-%m-%d", coerce=False)