@jreback 2015-07-18
The intention of that test behavior is that yes you will ignore errors (and you are not coercing), BUT, since you specified the format it is still an invalid date. 

so you bring up the case that it is doing different things depending on whether a format is passed or not.

rereading those comments in #10154. I think it is just a choice. Ok, so let's make them the same. I'll reverse myself and now say that all 4 of these should raise. @chris-b1 would you make that change.

cc @jorisvandenbossche 
cc @vincentdavis