@chris-b1 2015-07-19
%Y format with integer data doesn't work on iso fast code (causing Stata test problems) so I've excluded from this and added a new test.

I also changed the error handling in `to_datetime` to always raise on out of bound / unparse-able date unless `coerce=True`.  It ends up breaking quite a tests that expected the old behavior, just wanted to double check that's desired?  Thanks.