@jorisvandenbossche 2015-07-19
Rereading the comments in the other issue: https://github.com/pydata/pandas/issues/10154#issuecomment-106484971 and the two comments below. Based on that, I would say we agreed that with `errors='ignore'`, no error should be raised (although it is not really clear what we decided). But I did not see that the actual PR then did something else. 
I think it is pretty clear: if you give `errors='ignore'`, it should not raise an error. But as I said over there, I don't really like this default behaviour.