@chris-b1 2015-07-19
I think that makes sense.   One corner case that comes up - what should happen if the data passed doesn't match the format, but can all be parsed.  E.g.

    pd.to_datetime(['2/28/2014', '2014-01-01'], format='%m/%d/%Y')
    pd.to_datetime(['2/28/2014', '2014-01-01'], format='%m-%d-%Y')

I'm not sure, but it's probably least surprising if it fails? That's what the latest commit does - adding an additional parameter to `tslib.array_to_datetime` to only parse iso8601 strings.  I wasn't too sure on the cleanest way to do that, but here's a first attempt that at least seems to work.