@jreback 2015-07-19
@chris-b1 there are already a lot of paths thru this code, so getting very hard to grok. Any simplication would be appreciated.