@jorisvandenbossche 2015-07-19
After second thought, I am not sure anymore if this is a good idea (ignoring instead of raising), sorry ..

In any case, this is a backwards incompatible change, as it has been like that (that format raises instead of ignoring by default when it does not follow the format). It was just not correctly stated in the docstring. 
And if we do such a change, I would rather do it the other way around (change default of ignore to raise). But that is another discussion, and we should see if this change is not too breaking for people.

But maybe we can open a new issue for the discussion about the default behaviour, and keep this PR only for the perf improvement?