@sinhrks 2015-07-20
I think ``_string_to_dts`` should handle ``require_iso8601`` kw for simplification. Raise ``DateParseError`` when it is ``True``, and ``ValueError`` otherwise. ``array_to_datetime``, can pass through ``DateParseError``, but catch ``ValueError``. 

  - https://github.com/pydata/pandas/blob/master/pandas/src/datetime.pxd#L127

For ``errors`` kw, I don't have clear opinion for now.