@chris-b1 2015-07-20
@sinhrks is it possible to raise/define an custom exception from a .pxd?  Sorry, running into the edge of my Cython knowledge.