@sinhrks 2015-07-20
@chris-b1 Can use cdef (and convert it to normal python object when catch). But it does not simplify as I first thought... Please forget, and I'll do separately if I find better idea.