@jreback 2015-07-20
@chris-b1 this looks ok. pls squash; ping when green.

I opened #10636 to address the changing of the default, which is prob a good idea for 0.17.0

@chris-b1 if you'd like to tackle would be gr8!