@chris-b1 2015-07-20
@jreback - should be good to merge.