@jreback 2015-07-20
@chris-b1 thanks!

on #10636 we'll have to put the new behavior (that was slightly changed here).